+++ src/components/Form.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Form() {
+    return(
+        <React.Fragment>
+            
+        </React.Fragment>
+    );
+}
+
+export default Form;
